requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
faiss-cpu
sentence-transformers
requests
scikit-learn
